update auth and display user details in dashboard
